+++ examples/sweep/entrypoint.py
@@ -1,3 +1,11 @@
+"""
+This is script that will run on Beaker as the Docker image's "entrypoint".
+
+All it does is write out a simple JSON file with a random number in it to
+the experiment's result directory. This is just meant to simulate the results
+of a training/evaluation pipeline.
+"""
+
 import json
 import random
 import sys